Move hiding hydration warnings to entry-client.tsx
