+++ site/src/root.tsx
@@ -1,7 +1,6 @@
 // @refresh reload
-import { onMount, Suspense } from "solid-js";
+import { Suspense } from "solid-js";
 import {
-  A,
   Body,
   ErrorBoundary,
   FileRoutes,
@@ -14,7 +13,6 @@ import {
   Title,
 } from "solid-start";
 import { isServer } from "solid-js/web";
-import { debounce } from "@solid-primitives/scheduled";
 
 import "./root.scss";
 
@@ -29,24 +27,6 @@ if (isServer) {
 }
 
 export default function Root() {
-  // Primitives/Table.tsx produces a lot of hydration warnings in development mode.
-  if (import.meta.env.MODE === "development" && !isServer) {
-    const keys: string[] = [];
-    const cw = console.warn;
-    console.warn = (...args) => {
-      if (args[0] === "Unable to find DOM nodes for hydration key:") {
-        keys.push(args[1]);
-        logStoredWarnings();
-      } else cw(...args);
-    };
-    const logStoredWarnings = debounce(() => {
-      console.groupCollapsed(`There were ${keys.length} hydration warnings.`);
-      keys.forEach(key => cw("Unable to find DOM nodes for hydration key:", key));
-      console.groupEnd();
-      keys.length = 0;
-    }, 1000);
-  }
-
   return (
     <Html lang="en" data-html>
       <Head>
